chore: add skeleton for nfa list
