+++ ui/src/views/explore/explore-list/explore-list.fragment.tsx
@@ -1,12 +1,12 @@
 import { Explore } from '../explore.context';
-import { NFAListFragment } from './nfa-list';
+import { NFAsContainerFragment } from './nfa-list';
 import { NFASearchFragment } from './nfa-search.fragment';
 
 export const ExploreListFragment: React.FC = () => {
   return (
     <Explore.Provider>
       <NFASearchFragment />
-      <NFAListFragment />
+      <NFAsContainerFragment />
     </Explore.Provider>
   );
 };
