+++ ui/src/views/explore/explore-list/nfa-list/index.ts
@@ -1,1 +1,1 @@
-export * from './nfa-list.fragment';
+export * from './nfas-container.fragment';
