+++ ui/src/views/explore/explore-list/nfa-list/nfa-list.fragment.tsx
@@ -1,85 +1,58 @@
-import { useQuery } from '@apollo/client';
-import { useEffect } from 'react';
+import { Text } from '@/components';
+import { lastNFAsPaginatedQuery } from '@/graphclient';
 
-import { NFACard, NFACardSkeleton } from '@/components';
-import { lastNFAsPaginatedDocument } from '@/graphclient';
-import { useWindowScrollEnd } from '@/hooks';
-
-import { Explore } from '../../explore.context';
-import { NFAsListView } from './nfa-grid.fragment';
 import { NFAListFragmentStyles as S } from './nfa-list.styles';
+import { NFARow } from './nfa-row.fragment';
+import { NFARowSkeletonFragment } from './nfa-row.skeleton';
 
-const pageSize = 10; //Set this size to test pagination
-
-const LoadingSkeletons: React.FC = () => (
+const LoadingListSkeleton: React.FC = () => (
   <>
-    <NFACardSkeleton />
-    <NFACardSkeleton />
-    <NFACardSkeleton />
-    <NFACardSkeleton />
-    <NFACardSkeleton />
-    <NFACardSkeleton />
+    <NFARowSkeletonFragment />
+    <NFARowSkeletonFragment />
+    <NFARowSkeletonFragment />
   </>
 );
 
-export const NFAListFragment: React.FC = () => {
-  const {
-    endReached,
-    orderBy,
-    orderDirection,
-    pageNumber,
-    search,
-    nfaView,
-    setEndReached,
-    setPageNumber,
-  } = Explore.useContext();
+type NFAListFragmentProps = {
+  tokens: Array<lastNFAsPaginatedQuery['tokens'][0]>;
+  isLoading: boolean;
+};
 
-  const {
-    data: { tokens } = { tokens: [] },
-    loading: isLoading,
-    error: queryError,
-  } = useQuery(lastNFAsPaginatedDocument, {
-    fetchPolicy: 'cache-and-network',
-    variables: {
-      pageSize,
-      searchValue: search,
-      orderBy,
-      orderDirection,
-      skip: pageNumber * pageSize, //skip is for the pagination
-    },
-    onCompleted: (data) => {
-      if (data.tokens.length - tokens.length < pageSize) setEndReached(true);
-    },
-  });
+export const NFAListFragment: React.FC<NFAListFragmentProps> = ({
+  tokens,
+  isLoading,
+}: NFAListFragmentProps) => {
+  return (
+    <S.Table.Container>
+      <S.Table.Root>
+        <colgroup>
+          <col span={1} />
+          <col span={1} />
+          <col span={1} />
+        </colgroup>
+        <S.Table.Head>
+          <S.Table.Row>
+            <S.Table.Data>NAME</S.Table.Data>
+            <S.Table.Data># HOSTED</S.Table.Data>
+            <S.Table.Data>Owner</S.Table.Data>
+          </S.Table.Row>
+        </S.Table.Head>
+        <S.Table.Body>
+          {tokens.map((token) => (
+            <NFARow token={token} key={token.id} />
+          ))}
 
-  useEffect(() => {
-    // Update page number when there are cached tokens
-    setPageNumber(Math.ceil(tokens.length / pageSize));
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, []);
+          {isLoading && <LoadingListSkeleton />}
 
-  useWindowScrollEnd(() => {
-    if (isLoading || endReached || queryError) return;
-    setPageNumber(pageNumber + 1);
-  });
-
-  if (queryError) return <div>Error</div>; //TODO handle error
-
-  if (nfaView === 'grid')
-    return (
-      <S.Container>
-        {tokens.map((token) => (
-          <NFACard data={token} key={token.id} />
-        ))}
-
-        {isLoading && <LoadingSkeletons />}
-
-        {!isLoading && tokens.length === 0 && (
-          <S.EmptyMessage>Nothing found.</S.EmptyMessage>
-        )}
-      </S.Container>
-    );
-  else {
-    return <NFAsListView accessPoints={tokens} />;
-  }
+          {!isLoading && tokens.length === 0 && (
+            <S.Table.Row>
+              <S.Table.Data align="center" colSpan={5}>
+                <Text>No results</Text>
+              </S.Table.Data>
+            </S.Table.Row>
+          )}
+        </S.Table.Body>
+      </S.Table.Root>
+    </S.Table.Container>
+  );
 };
